# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Voitity API",
        "description": "API documentation for Voitity",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payments/wompi/checkout": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a Wompi checkout for a subscription plan",
                "operationId": "ddfbcbfce646e936ba2a56d8692c6afa",
                "responses": {
                    "201": {
                        "description": "Wompi checkout created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": [
                            "payments:create"
                        ]
                    }
                ]
            }
        },
        "/api/payments/{paymentOrder}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get a payment order",
                "operationId": "dd3389c3a2009d8c435263e307c7d6f9",
                "responses": {
                    "200": {
                        "description": "Payment order retrieved successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": [
                            "payments:read"
                        ]
                    }
                ]
            }
        },
...
        "/api/subscription/limits": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get active subscription limits",
                "operationId": "22be609c4045c43bf19faaf25a39dc90",
                "responses": {
                    "200": {
                        "description": "Subscription limits retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription limits retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "subscription": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "user_id": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "plan": {
                                                            "type": "string",
                                                            "example": "starter"
                                                        },
                                                        "plan_name": {
                                                            "type": "string",
                                                            "example": "Starter"
                                                        },
                                                        "price_usd": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 8
                                                        },
                                                        "currency": {
                                                            "type": "string",
                                                            "example": "USD"
                                                        },
                                                        "interval": {
                                                            "type": "string",
                                                            "example": "monthly"
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "first"
                                                        },
                                                        "active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "started_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "renews_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "limits": {
                                                    "type": "object"
                                                },
                                                "usage": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Active subscription not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": [
                            "subscription-limits:read"
                        ]
                    }
                ]
            }
        },
...
    },
...
}